Moved the extended piece to a subdirectory of existing object_detection_2d example
