+++ examples/object_detection_2d/object_detection_2d/extended/seed_test_suite.py
@@ -23,17 +23,17 @@ from typing import Tuple
 
 import pandas as pd
 import s3fs
-from constants import DATASET
-from constants import S3_ANNOTATION_FILE_PATH
-from constants import S3_IMAGE_LOCATION
-from constants import TEST_SUITE_DESCRIPTION
-from constants import TRANSPORTATION_LABELS
-from constants import WORKFLOW
-from object_detection_2d_extend.workflow import ExtendedBoundingBox
-from object_detection_2d_extend.workflow import GroundTruth
-from object_detection_2d_extend.workflow import TestCase
-from object_detection_2d_extend.workflow import TestSample
-from object_detection_2d_extend.workflow import TestSuite
+from object_detection_2d.constants import DATASET
+from object_detection_2d.constants import S3_ANNOTATION_FILE_PATH
+from object_detection_2d.constants import S3_IMAGE_LOCATION
+from object_detection_2d.constants import TEST_SUITE_DESCRIPTION
+from object_detection_2d.constants import TRANSPORTATION_LABELS
+from object_detection_2d.constants import WORKFLOW
+from object_detection_2d.extended.workflow import ExtendedBoundingBox
+from object_detection_2d.extended.workflow import GroundTruth
+from object_detection_2d.extended.workflow import TestCase
+from object_detection_2d.extended.workflow import TestSample
+from object_detection_2d.extended.workflow import TestSuite
 
 import kolena
 from kolena.workflow.annotation import BoundingBox
