--- examples/object_detection_2d_extend/pyproject.toml
@@ -1,22 +1,0 @@
-[tool.poetry]
-name = "object_detection_2d_extend"
-version = "0.1.0"
-description = "Example Kolena integration for 2D object detection with extended `GroundTruth`"
-authors = ["Kolena Engineering <eng@kolena.io>"]
-license = "Apache-2.0"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.12"
-kolena = {extras = ["metrics"], version = ">=0.76.0,<1"}
-s3fs = "^2023.5.0"
-fsspec = "^2023.5.0"
-pandera = ">=0.9.0,<0.16"
-
-[tool.poetry.group.dev.dependencies]
-pre-commit = "^2.17"
-pytest = "^7"
-pytest-depends = "^1.0.1"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
